+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Don't retry requests to SEKOIA.IO with a status code of 4xx and ignore 409 errors
 
+### Fixed
+
+- Specify timeout to HTTP requests
+
 ## [1.1.0b1] - 2023-03-16
 
 ### Added